+++ InfinityHammer/InfinityHammer.cs
@@ -11,13 +11,13 @@ namespace InfinityHammer;
 [BepInPlugin(GUID, NAME, VERSION)]
 [BepInDependency("com.rolopogo.gizmo.comfy", BepInDependency.DependencyFlags.SoftDependency)]
 [BepInDependency("bruce.valheim.comfymods.gizmo", BepInDependency.DependencyFlags.SoftDependency)]
-[BepInDependency("server_devcommands", "1.99")]
-[BepInDependency("world_edit_commands", "1.67")]
+[BepInDependency("server_devcommands", "1.109")]
+[BepInDependency("world_edit_commands", "1.74")]
 public class InfinityHammer : BaseUnityPlugin
 {
   public const string GUID = "infinity_hammer";
   public const string NAME = "Infinity Hammer";
-  public const string VERSION = "1.79.10";
+  public const string VERSION = "1.80";
   public static bool StructureTweaks = false;
 #nullable disable
   public static ConfigWrapper Wrapper;