+++ InfinityHammer/selection/BaseSelection.cs
@@ -19,7 +19,7 @@ public class BaseSelection
   public virtual float DungeonRoomSnapMultiplier => 1f;
   public virtual float SnapMultiplier => 1f;
   public virtual float MaxPlaceDistance(float value) => Configuration.Range > 0f ? Configuration.Range : value;
-  public Piece GetSelectedPiece() => SelectedPrefab ? SelectedPrefab.GetComponent<Piece>() : null!;
+  public virtual Piece GetSelectedPiece() => SelectedPrefab ? SelectedPrefab.GetComponent<Piece>() : null!;
   public virtual void Destroy() => Object.Destroy(SelectedPrefab);
 
   public virtual DataEntry? GetData(int index = 0) => null;